backend/requirements.txt
aiohttp==3.9.3
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
pydantic==2.6.3
pydantic-settings==2.2.1
python-dateutil==2.9.0
aiosqlite==0.20.0
email-validator==2.1.0